cakefactory#6

+++ src/main/java/com/factory/cake/domain/dto/BasketLineDTO.java
@@ -1,5 +1,7 @@
 package com.factory.cake.domain.dto;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -11,8 +13,9 @@ public class BasketLineDTO {
 	
     private String id;
     private String name;
+//    private String descritpion;
     private int quantity;
-	private float lineTotalPrice;
+	private BigDecimal lineTotalPrice;
 
 
 }
