+++ src/main/java/com/factory/cake/domain/dto/ItemDTO.java
@@ -1,5 +1,7 @@
 package com.factory.cake.domain.dto;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -11,7 +13,7 @@ public class ItemDTO  {
 
 	private String id;
 	private String name;
-	private float price;
+	private BigDecimal price;
 	private String description;
 	private String image;
 }
